+++ CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Full reentrancy, supporting multiple concurrent evaluations of the same `Expression` object
  - New multithreading model independent of Node.js/libuv
  - Support writing into a preallocated array for `Expression.prototype.map`
+ - Support passing `TypedArray` subarrays (`TypedArray.prototype.subarray`)
  - `Expression.prototype.toString()`
 
 ### [1.0.1] 2022-01-03